Add standalone postgres compose
This compose will be used for new persistant setup instructions. It is a
standalone file and doesn't need any additional files to be mounted into the
containers. It is a complete single file (not multiple files merged) and
doesn't contain anything unecessary. Future docs will build upon it for adding
additional collectors, etc.

Signed-off-by: Jeff Mendoza <jlm@jlm.name>